+++ R/elo.colley.R
@@ -1,5 +1,3 @@
-#' \code{elo.colley}
-#'
 #' Compute a Colley matrix model for a matchup.
 #'
 #' @inheritParams elo.glm
@@ -33,6 +31,7 @@ NULL
 elo.colley <- function(formula, data, family = "binomial", weights, na.action, subset, k = 1, ..., running = FALSE, skip = 0)
 {
   Call <- match.call()
+  Call <- Call[c(1, match(c("formula", "data", "weights", "subset", "na.action", "k"), names(Call), nomatch = 0))]
   Call[[1L]] <- quote(elo::elo.model.frame)
   Call$required.vars <- c("wins", "elos", "group", "neutral", "weights", "k")
   if(is.null(Call$k)) Call$k <- 1
@@ -82,8 +81,8 @@ elo.colley <- function(formula, data, family = "binomial", weights, na.action, s
 
     for(i in setdiff(seq_len(max(grp2)), seq_len(skip)))
     {
-      if(i == 1) next
-      sbst <- grp2 %in% 1:(i-1)
+      if(i == 0) next
+      sbst <- grp2 %in% 0:(i-1)
       dat.tmp <- dat
       dat.tmp$winsA <- dat.tmp$winsA[sbst]
       dat.tmp$k <- dat.tmp$k[sbst, , drop = FALSE]
@@ -94,6 +93,7 @@ elo.colley <- function(formula, data, family = "binomial", weights, na.action, s
       colley <- do.call(eloColley, dat.tmp)
       vec <- stats::.lm.fit(colley[[1]], colley[[2]])$coefficients
       vec <- stats::setNames(vec / sum(vec), all.teams)
+      vec[colley[[3]] == 0] <- NA
       difference <- mean_vec_subset_matrix(vec, dat$teamA+1) - mean_vec_subset_matrix(vec, dat$teamB+1)
 
       # tmpfit <- stats::glm(dat$winsA ~ difference, subset = sbst, family = "binomial")
@@ -101,9 +101,10 @@ elo.colley <- function(formula, data, family = "binomial", weights, na.action, s
 
       coeff <- stats::glm.fit(cbind(difference, adj)[sbst, , drop=FALSE],
                               dat.tmp$winsA, family = mc.glm$family, control = mc.glm$control)$coefficients
-      ftd[grp2 == i] <- apply(cbind(difference, adj)[grp2 == i, , drop=FALSE], 1, function(x) sum(x * coeff, na.rm = TRUE))
+      ftd[grp2 == i] <- apply(cbind(difference, adj)[grp2 == i, , drop=FALSE], 1, mult_na_coef, coeff = coeff)
     }
     out$running.values <- mc.glm$family$linkinv(ftd)
+    attr(out$running.values, "group") <- grp2
   }
 
   structure(out, class = c(if(running) "elo.running", "elo.colley"))