+++ R/elo.prob.R
@@ -1,4 +1,4 @@
-#' Elo functions
+#' Elo probability
 #'
 #' Calculate the probability that team A beats team B. This is vectorized.
 #'
@@ -72,8 +72,22 @@ elo.prob.default <- function(elo.A, elo.B, ..., elos = NULL, adjust.A = 0, adjus
 elo.prob.formula <- function(formula, data, na.action, subset, ..., elos = NULL)
 {
   Call <- match.call()
+  Call <- Call[c(1, match(c("formula", "data", "subset", "na.action"), names(Call), nomatch = 0))]
   Call[[1L]] <- quote(elo::elo.model.frame)
   mf <- eval(Call, parent.frame())
 
   elo.prob(mf$elo.A, mf$elo.B, ..., adjust.A = mf$adj.A, adjust.B = mf$adj.B, elos = elos)
+}
+
+#' @rdname elo.prob
+#' @export
+elo.prob.elo.multiteam.matrix <- function(elo.A, ..., elos = NULL)
+{
+  elo.A <- unclass(elo.A)
+  all.teams <- sort(unique(as.vector(elo.A)))
+  elos <- check_named_elos(elos, all.teams)
+
+  elo.A <- matrix(elos[elo.A], nrow = nrow(elo.A))
+  out <- 10^(elo.A/400.0)
+  out / rowSums(out, na.rm = TRUE)
 }