+++ R/elo.run.helpers.R
@@ -1,7 +1,8 @@
 
 #' Helper functions for \code{elo.run}
 #'
-#' \code{as.matrix} converts an Elo object into a matrix of running Elos.
+#' \code{as.matrix} converts an Elo object into a matrix of running Elos. These are the Elos at the time of grouping,
+#'   but before any regression takes place.
 #'
 #' \code{as.data.frame} converts the \code{"elos"} component of an object
 #'   from \code{\link{elo.run}} into a data.frame.
@@ -10,7 +11,6 @@
 #'
 #' @param x An object of class \code{"elo.run"} or class \code{"elo.run.regressed"}.
 #' @param ... Other arguments (Not in use at this time).
-#' @param group A grouping vector, telling which rows to output in the matrix.
 #' @param regressed Logical, denoting whether to use the post-regressed (\code{TRUE}) or
 #'   pre-regressed (\code{FALSE}) final Elos. Note that \code{TRUE} only makes sense when the
 #'   final Elos were regressed one last time (i.e., if the last element of the \code{regress()})
@@ -29,9 +29,9 @@ NULL
 
 #' @rdname elo.run.helpers
 #' @export
-as.matrix.elo.run <- function(x, ..., group = x$group)
+as.matrix.elo.run <- function(x, ...)
 {
-  group <- check_as_matrix(x, group)
+  group <- check_as_matrix(x, x$group)
   out <- eloRunAsMatrix(x$elos, x$initial.elos, group)
   colnames(out) <- x$teams
   out
@@ -39,9 +39,9 @@ as.matrix.elo.run <- function(x, ..., group = x$group)
 
 #' @rdname elo.run.helpers
 #' @export
-as.matrix.elo.run.regressed <- function(x, ..., group = x$group)
+as.matrix.elo.run.regressed <- function(x, ...)
 {
-  group <- check_as_matrix(x, group, regr = TRUE)
+  group <- check_as_matrix(x, x$group, regr = TRUE)
   out <- eloRunRegressedAsMatrix(x$elos, x$initial.elos, x$elos.regressed,
                                  check_group_regress(x$regress),
                                  group)
@@ -54,9 +54,14 @@ as.matrix.elo.run.regressed <- function(x, ..., group = x$group)
 as.data.frame.elo.run <- function(x, ...)
 {
   out <- as.data.frame(x$elos)
-  colnames(out) <- c("team.A", "team.B", "p.A", "wins.A", "update.A", "update.B", "elo.A", "elo.B")
-  out$team.A <- factor(out$team.A, levels = seq_along(x$teams), labels = x$teams)
-  out$team.B <- factor(out$team.B, levels = seq_along(x$teams), labels = x$teams)
+  nm.a <- if(x$n.players[1] > 1) paste0(".", seq_len(x$n.players[1])) else ""
+  nm.b <- if(x$n.players[2] > 1) paste0(".", seq_len(x$n.players[2])) else ""
+
+  colnames(out) <- c(paste0("team.A", nm.a), paste0("team.B", nm.b),
+                     "p.A", "wins.A", "update.A", "update.B",
+                     paste0("elo.A", nm.a), paste0("elo.B", nm.b))
+  out[paste0("team.A", nm.a)] <- lapply(out[paste0("team.A", nm.a)], factor, levels = seq_along(x$teams), labels = x$teams)
+  out[paste0("team.B", nm.b)] <- lapply(out[paste0("team.B", nm.b)], factor, levels = seq_along(x$teams), labels = x$teams)
   out
 }
 
@@ -82,7 +87,7 @@ final.elos.elo.run <- function(x, ...)
 final.elos.elo.run.regressed <- function(x, regressed = FALSE, ...)
 {
   if(regressed && !utils::tail(check_group_regress(x$regress), 1))
-    warning("'regressed = TRUE' only makes sense if the final Elos are regressed after the final game.")
+    stop("'regressed = TRUE' only makes sense if the final Elos are regressed after the final game.")
 
   if(!regressed) return(NextMethod())