+++ R/elo.update.R
@@ -1,5 +1,5 @@
 
-#' Elo functions
+#' Elo updates
 #'
 #' Calculate the update value for a given Elo matchup. This is used in
 #' \code{\link{elo.calc}}, which reports the post-update Elo values. This is vectorized.
@@ -38,6 +38,7 @@ elo.update.default <- function(wins.A, elo.A, elo.B, k, ..., adjust.A = 0, adjus
 elo.update.formula <- function(formula, data, na.action, subset, k = NULL, ...)
 {
   Call <- match.call()
+  Call <- Call[c(1, match(c("formula", "data", "subset", "na.action", "k"), names(Call), nomatch = 0))]
   Call[[1L]] <- quote(elo::elo.model.frame)
   Call$required.vars <- c("wins", "elos", "k")
   mf <- eval(Call, parent.frame())